pyproject.toml
[project]
name = "django-order-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.6",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-mail-templated>=2.6.5",
    "django-redis>=5.4.0",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "psycopg[binary]>=3.2.4",
    "redis>=5.2.1",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=5.0.1",
    "factory-boy>=3.3.3",
    "pytest-django>=4.10.0",
    "pytest-mock>=3.14.0",
]
